drm/vc4: Fix frame corruption when moving the cursor
Hi,

This series fixes a long standing issue with the VC4 driver when one was
moving the cursor on X11 along the edges of the monitor, if we had
overscan margins enabled.

The details are in the commit log of the last patch, but the main reason
was that moving along the edges with the overscan margins enabled
triggers a full blown commit, as opposed to an async commit. Since that
commit is on the cursor plane, it's treated as a legacy cursor update,
and won't wait for vblank, so it's possible to queue multiple commit
between vblank.

Now, the composition happens in the HVS, and the HVS has a series of
descriptors stored in an internal RAM, one for each plane. Allocations
in that RAM are tied to the CRTC state, and freed when that state is
destroyed. That internal RAM is also used by the HVS to store some
internal context while it's generating a frame.

If we get multiple commits between vblanks, chances are that the RAM
entries used by one of the first commit is going to be freed and reused
by a later commit, which will then overwrite the content of the earlier
entries, erasing the context in the process and corrupting the frame.

We've tested multiple solutions, but the one that seem to work without
any cons is to defer the deallocation of RAM entries to the next vblank
after the CRTC state has been freed.

In the solutions we've considered and tested, removing the legacy cursor
updates entirely has done in:
https://lore.kernel.org/all/20220331152021.2671937-1-daniel.vetter@ffwll.ch/

did indeed improve the situation, but we were still getting some similar
bugs when running Kodi, as reported here:
https://github.com/raspberrypi/linux/pull/5327

That patch makes them go away entirely, and since the legacy cursor
updates are still there after a bit less than a year, I believe we
should merge that patch.

Let me know what you think,
Maxime

To: Emma Anholt <emma@anholt.net>
To: Maxime Ripard <mripard@kernel.org>
To: David Airlie <airlied@gmail.com>
To: Daniel Vetter <daniel@ffwll.ch>
Cc: Thomas Zimmermann <tzimmermann@suse.de>
Cc: Maarten Lankhorst <maarten.lankhorst@linux.intel.com>
Cc: Dave Stevenson <dave.stevenson@raspberrypi.com>
Cc: Phil Elwell <phil@raspberrypi.com>
Cc: Dom Cobley <dom@raspberrypi.com>
Cc: dri-devel@lists.freedesktop.org
Cc: linux-kernel@vger.kernel.org
Signed-off-by: Maxime Ripard <maxime@cerno.tech>

---
Changes in v3:
- EDITME: describe what is new in this series revision.
- EDITME: use bulletpoints and terse descriptions.
- Link to v2: https://lore.kernel.org/r/20230127-rpi-cursor-corruption-v2-1-1f97bd00df01@cerno.tech

Changes in v2:
- Rebased against current next
- Fixed a kunit breakage
- Link to v1: https://lore.kernel.org/dri-devel/20220221134155.125447-1-maxime@cerno.tech/

--- b4-submit-tracking ---
# This section is used internally by b4 prep for tracking purposes.
{
  "series": {
    "revision": 3,
    "change-id": "20230127-rpi-cursor-corruption-c98fc609ab18",
    "base-branch": null,
    "prefixes": [],
    "history": {
      "v1": [
        "20220221134155.125447-1-maxime@cerno.tech"
      ],
      "v2": [
        "20230127-rpi-cursor-corruption-v2-1-1f97bd00df01@cerno.tech"
      ]
    }
  }
}